Introduce PairData, dedicated to instance pairs, as a subclass of FeatureData
The PairData class makes explicit that a feature array contains features of paired instances, and the semantics of the feature array itself. It has information on the number of instances involved, that otherwise has to be derived from context information, and it has attributes to get the trace/reference side of the pairs.