Allow arguments to pipeline steps in LR system YAML definitions
To be able to parse extra arguments of steps in a defined LR system, the
class may now be referenced using the `method:` key. If no such key is
defined, a default will be choosen. For most steps this will be the
default 'pipeline' method.

This patch makes it easier to separate, e.g. the CSV writing steps.