Combine aggregated metrics+visualization to single 'output' key in YAML (#84)
* aggregation: rename YAML key 'aggregation' to 'output'

This renaming better represents its meaning and allows a CSV containing the metrics to be written to the output folder. This step also prepares for the merge of visualization output now under a 'visualization' key to be contained in the 'output' key in a next patch.

* aggregation: integrate visualization functions in YAML 'output' key

The config parser which parses the `output` section, now expects an `Aggregation` class instead of a function, which is why the visualization functions were replaced by `AggregatePlot` class references instead.

* testing: added unit test that runs all registered output aggregation methods

---------

Co-authored-by: wauter <20113294+wowtor@users.noreply.github.com>
Co-authored-by: Pim Meulensteen <pimmeulensteen+git@gmail.com>